test: move instancetype unit tests to its own package
