+++ pkg/handlers/aws/mutation/instancetype/inject_suite_test.go
@@ -1,0 +1,13 @@
+package instancetype
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestInstanceTypePatch(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "InstanceType patches for ControlPlane and Workers suite")
+}
